requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
pydantic-settings==2.1.0
typer==0.9.0
pydantic==2.6.1
pydantic[email]
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
python-multipart==0.0.7
aiosmtplib==3.0.1
python-jose==3.3.0
sqlmodel==0.0.18
psycopg2-binary==2.9.9
asyncpg==0.29.0
